askfaithbot-rag-lambda/requirements.txt
boto3
botocore
python-dotenv
langchain
langchain-aws
langchain-community
gpt4all
faiss-cpu
